extracted createSolutionPageContext to a seperate file and removed contextCreator file after refactor
